requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
slowapi==0.1.9
sendgrid==6.11.0
python-dotenv==1.0.0
email-validator